worker/requirements.txt
celery==5.3.6
redis==5.0.7
pymupdf==1.24.5
pillow==10.3.0
requests==2.32.3
prometheus-client==0.20.0
openpyxl==3.1.4
